Refractor file structures and improve tests
