+++ src/exceptions/stumped_lovecraft_exception.py
