+++ src/model/lovecraft_model.py
